requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorlog==6.8.2
confection==0.1.4
cymem==2.0.8
databricks-cli==0.18.0
databricks-sql-connector==2.5.2
dataclasses-json==0.6.1
distro==1.9.0
Django==4.1.12
docker==7.1.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.14.0
frozenlist==1.4.0
fsspec==2024.5.0
fuzzywuzzy==0.18.0
greenlet==3.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.312
langcodes==3.4.0
langsmith==0.0.43
language_data==1.2.0
Levenshtein==0.25.1
lz4==4.3.2
Mako==1.2.4
marisa-trie==1.1.1
MarkupSafe==2.1.5
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.30.2
openpyxl==3.1.2
packaging==24
pandas==1.5.3
pillow==10.3.0
preshed==3.0.9
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyarrow==13.0.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyspark==3.5.0
python-dateutil==2.8.2
python-Levenshtein==0.25.1
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.9.1
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.7.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.49
sqlparse==0.4.4
srsly==2.4.8
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
textblob==0.18.0.post0
thinc==8.2.3
threadpoolctl==3.5.0
thrift==0.16.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.1
typer==0.9.4
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
whitenoise==6.5.0
yarl==1.9.2